+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+esriLoader defaults to loading JSAPI v4.2. Docs site uses JSAPI v4.2. [#317](https://github.com/Esri/angular-esri-map/issues/317)
+
 Added support for Angular v1.6.0 and greater due to Angular's breaking changes with pre-assigning bindings on controller instances. [#314](https://github.com/Esri/angular-esri-map/issues/314)
 
 ### Maintenance